Add logging for OSError exceptions in the Signal transport class
signal-cli has some stability issues, this adds extra
logging so its easier to keep an eye on it.

With the default Django logging configuration, when OSError
is raised, Django will send an error report to ADMINS.

If sentry_sdk is installed and configured, the exception
will also show up in Sentry.